Clean up the code. Make user in VideoPager clickable and redirect to user profile. Add delayed worker executed every day which removes orphaned posts in room database. Fix some minor bugs related to setting profile information (followers, followings) to null when loading posts.

+++ app/src/main/java/live/foodclub/utils/composables/videoPager/VideoPager.kt
@@ -177,7 +177,7 @@ fun VideoPager(
                         onVideoGoBackground = { pauseButtonVisibility = false }
                     )
 
-                    if (currentVideo.authorDetails.userId == state.browsingUserId) {
+                    if (currentVideo.authorDetails.userId == state.browsingUserId && !state.isHomeView) {
                         DeleteButton(
                             alignment = Alignment.TopEnd,
                             onDeleteClicked = { deleteDialog.value = true }
